Increase the scale of the demo by a fator of 10.
While this makes for a *huge* tulip (1.6-metre-wide pot!), it demonstrates that
the quantisation issues with the earlier demo occured merely due to the small
scale. This may yet be something to look into, though, after we see what
happens in real-world rendering. Hulls generally won't be used for small-scale
things, though.