Switch to point sprites for glbrush splotches.
This *massively* reduces the amount of memory bandwidth used for the
splotches. A quad splotch required 4 vertices of 5 floats plus six indices, for
a total of 92 bytes per splotch. With point sprites, one vertex of 6 floats and
so indices, for only 24 bytes. It also produces lower computational load on the
CPU, and very marginally less on the vertex processor in the GPU.

The notes on glbrush.c:200 suggest this might be a bit finicky on other GL
implementations. We'll need to keep that in mind when running elsewhere.