Vendor in Lua 5.2.3.
Bringing this in because we need to make Lua use `long long` for numbers
instead of `double` no matter what. While here, also going to bring in a
number of the "power patches" to make the language nicer (manually, because
virtually none of them apply cleanly...).

Doing this also has the nice side-benefit that every build will be using
exactly the same Lua version.