Paint overlay now looks like actual paint!
The paint texture itself still needs a little work. The way extra data
(like rotation) is communicated will also need to be redisigned, since
we can't afford the performance pentalty of reading from three textures.
Also, the concept of pass-through (not implemented; originally intended
for the terrain lines) needs to simply be implemented as a second pass
over the paint overlay, since it would still result in big ugly
splotches of black occasionally.

Still gets lots of shimmering due to the lines in the terrain. This will
require using simpler colour variance when rendering the terrabuff, and
maybe using real textures for terrain near the camera.