Initial work on llua bindings.
Custom code generator, somewhat unfortunately. But SWIG isn't safe (eg,
makes dangling pointers possible if they're exported, and in any case
there's a lot of machinery we don't really want). tolua requires a
runtime module, which of course would need to be patched to support
Llua, so we'd then have our own code generator/runtime anyway, and I
don't think we really need all it does in any case. So sticking with the
simplest possible solution (in terms of runtime support and
capabilities) for now.