Allow the paint overlay to be rendered at reduced resolution.
There's still some issues with this. The paint overlay does a poor job of
filling the screen at lower resolutions, for some reason. The grahpte effect
also somehow ends up being an all-or-nothing-for-whole-pixel-column effect
somehow, being invisible unless it occupies the very top of the screen, in
which case it occupies everything on that X coordinate.

For now, committing this as-is for performance testing.