Fix assertion error in terrabuff_merge.
Ie, remove the assertion and work around the condition where it occurs.
This was only happening because our non-linear spacial projection
violates the terrabuff's assumptions when the camera is at extreme rx
angles.

Finally doing this now since it's rather annoying to have crashes that
occur when looking up when one is trying to do stuff with the sky.