Drop the voxel renderer.
Originally it was planned to be used for rigid structures and
grass/bushes. The former can be handled by the a second manifold
renderer which does not do subdivision (and it'll do a better job,
besides also being easier to use); the latter will instead be
dynamically rendered point/line graphs or something.